refactor: extract more functions from tests.cpp
