+++ src/utils/io_utils.h
@@ -4,33 +4,38 @@
 #include <string>
 #include <vector>
 #include <map>
+#include "json.hpp"
+
+using namespace std;
+using json = nlohmann::json;
 
 // Structure to hold reference sequence information
 struct Reference
 {
-    std::string name;
-    std::string sequence;
+    string name;
+    string sequence;
     double nrc = 0.0;
     double kld = 0.0;
     double compressionBits = 0.0;
 };
 
 // File reading functions
-std::string readMetagenomicSample(const std::string &filename);
-std::vector<Reference> readReferenceDatabase(const std::string &filename);
+string readMetagenomicSample(const string &filename);
+vector<Reference> readReferenceDatabase(const string &filename);
 
 // Configuration parsing functions
-bool parseConfigFile(const std::string &configFile, std::map<std::string, std::string> &configParams);
-bool stringToBool(const std::string &value);
+bool parseConfigFile(const string &configFile, map<string, string> &configParams);
+bool stringToBool(const string &value);
 
 // Results saving functions
-bool saveResultsToJson(const std::vector<std::pair<std::pair<int, double>, std::pair<std::vector<Reference>, double>>> &allResults,
-                       const std::string &outputFile);
-bool saveResultsToCsv(const std::vector<std::pair<std::pair<int, double>, std::pair<std::vector<Reference>, double>>> &allResults,
-                      const std::string &outputFile);
-bool saveAllResultsToJson(const std::vector<std::pair<std::pair<int, double>, std::pair<std::vector<Reference>, double>>> &allResults,
-                          const std::string &outputFile);
-bool saveAllResultsToCsv(const std::vector<std::pair<std::pair<int, double>, std::pair<std::vector<Reference>, double>>> &allResults,
-                         const std::string &outputFile);
+bool saveResults(const json& results, const string& timestampFile, const string& latestFile);
+bool saveResultsToJson(const vector<pair<pair<int, double>, pair<vector<Reference>, double>>> &allResults,
+                       const string &outputFile);
+bool saveResultsToCsv(const vector<pair<pair<int, double>, pair<vector<Reference>, double>>> &allResults,
+                      const string &outputFile);
+bool saveAllResultsToJson(const vector<pair<pair<int, double>, pair<vector<Reference>, double>>> &allResults,
+                          const string &outputFile);
+bool saveAllResultsToCsv(const vector<pair<pair<int, double>, pair<vector<Reference>, double>>> &allResults,
+                         const string &outputFile);
 
 #endif // IO_UTILS_H
